refactor: move all docker-related functions to new file
